third_party/WebKit/LayoutTests/inspector/console/console-proxy-expected.txt

CONSOLE MESSAGE: line 21: [object Proxy]
CONSOLE MESSAGE: line 23: [object Proxy]
Tests that console logging dumps proxy properly.

warning-note display: inline-block
info-note display: none
console-proxy.html:21 Proxy {boo: 42, foo: 43} console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element > selection fill > console-object-preview > name > object-value-number > name > object-value-number > object-state-note warning-note > object-state-note info-note > children
console-proxy.html:23 Proxy {boo: 42, foo: 43} console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element > selection fill > console-object-preview > name > object-value-number > name > object-value-number > object-state-note warning-note > object-state-note info-note > children
window.accessedGet = false
warning-note display: none
info-note display: inline-block
console-proxy.html:21 Proxy {boo: 42, foo: 43}boo: 42foo: 43__proto__: Object console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section expanded > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element expanded > selection fill > console-object-preview > name > object-value-number > name > object-value-number > object-state-note warning-note > object-state-note info-note > children expanded > selection fill > name > object-properties-section-separator > value object-value-number > children > selection fill > name > object-properties-section-separator > value object-value-number > children > parent > selection fill > name object-properties-section-dimmed > object-properties-section-separator > value object-value-object > children
console-proxy.html:23 Proxy {boo: 42, foo: 43}boo: 42foo: 43__proto__: Object console-message-text source-code > console-message-url webkit-html-resource-link > object-value-proxy source-code > console-view-object-properties-section expanded > tree-outline-disclosure tree-outline-disclosure-hide-overflow > tree-outline source-code object-properties-section > parent object-properties-section-root-element expanded > selection fill > console-object-preview > name > object-value-number > name > object-value-number > object-state-note warning-note > object-state-note info-note > children expanded > selection fill > name > object-properties-section-separator > value object-value-number > children > selection fill > name > object-properties-section-separator > value object-value-number > children > parent > selection fill > name object-properties-section-dimmed > object-properties-section-separator > value object-value-object > children
window.accessedGet = true
